pyproject.toml
[project]
name = "backend-as-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "fastapi>=0.116.1",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langchain-openapi>=0.1.1",
    "langchain-tavily>=0.2.11",
    "numpy>=2.3.3",
    "openai>=1.107.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "uuid>=1.30",
]
